backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
SQLAlchemy==2.0.25
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
google-generativeai==0.3.2
PyJWT==2.8.0
bcrypt==4.1.2
email-validator==2.1.0
aiosqlite==0.19.0
